requirements.txt
streamlit
boto3
langchain
langchain-aws
langchain-experimental
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
PyMuPDF
pypdf
openai
requests
python-dotenv
beautifulsoup4
pandas
numpy==1.26.4
pydantic
pydantic-settings
flask
flask-cors
werkzeug
gnews
pyjwt
bcrypt
streamlit-authenticator